Removed all occurance of decimal and converted to double.

+++ Application/Salvation.Core/Constants/BaseSpellData.cs
@@ -34,28 +34,28 @@ namespace Salvation.Core.Constants
         public double ManaCost { get; set; }
         // Casting range
         public double MaxRange { get; set; }
-        public decimal NumberOfDamageTargets { get; set; }
+        public double NumberOfDamageTargets { get; set; }
         // Base cast time in seconds before haste
-        public decimal BaseCastTime { get; set; }
+        public double BaseCastTime { get; set; }
         public bool IsCastTimeHasted { get; set; }
-        public decimal BaseCooldown { get; set; }
+        public double BaseCooldown { get; set; }
         public bool IsCooldownHasted { get; set; }
-        public decimal Duration { get; set; }
+        public double Duration { get; set; }
         // Base GCD in seconds before haste
-        public decimal Gcd { get; set; }
+        public double Gcd { get; set; }
 
         /// <summary>
         /// Spellpower coefficient for component #1
         /// </summary>
-        public decimal Coeff1 { get; set; }
+        public double Coeff1 { get; set; }
         /// <summary>
         /// Spellpower coefficient for component #2
         /// </summary>
-        public decimal Coeff2 { get; set; }
+        public double Coeff2 { get; set; }
         /// <summary>
         /// Spellpower coefficient for component #3
         /// </summary>
-        public decimal Coeff3 { get; set; }
+        public double Coeff3 { get; set; }
         // If mastery is triggered for the direct heal portion
         // TODO: Move this out to a HolyPriestSpellData inherited class
         public bool IsMasteryTriggered { get; set; }
